backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.23.2
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pydantic[email]>=2.4.2
pydantic-settings>=2.0.3
email-validator>=2.1.0
beautifulsoup4>=4.12.2
requests>=2.31.0
feedparser>=6.0.10
torch>=2.1.0 --index-url https://download.pytorch.org/whl/cpu
transformers>=4.35.0
nltk>=3.8.1
vaderSentiment>=3.3.2
apscheduler>=3.10.4
pandas>=2.1.2
scikit-learn>=1.3.2
google-genai>=0.5.0
httpx>=0.25.0
python-multipart>=0.0.6
